Improve styling
