+++ code/src/App.css
@@ -1,6 +1,10 @@
 main {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
+  max-width: 400px;
+  margin: 40px auto 0;
+}
+
+.card-wrap {
+  border: 1px solid black;
+  box-shadow: 8px 8px black;
+  margin: 0 15px 40px 15px;
 }
